Update stats when things changed.
issue #12031
